requirements.txt
Django==4.1.7
psycopg2
python-dotenv
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
